Broke out the template data methods and added local policy to template
